Move file to internal nearcache package
