+++ hazelcast/src/main/java/com/hazelcast/internal/nearcache/impl/NearCachingHook.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.client.cache.impl.nearcache;
+package com.hazelcast.internal.nearcache.impl;
 
 
 import com.hazelcast.internal.serialization.Data;
